[bg #162077005] Fix bug that was causing heroku to crash
