+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\__init__.py":[2,3,4,5,6,8,10,11,12,13,16,23],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\views\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\views\\product_views.py":[1,2,3,4,5,6,8,9,10,11,14,15,36,37,46,47,56,57,84,85,100,101,136,137,156,157,167,17,18,19,20,23,24,25,27,28,30,31,32,33,50,51,53,29,40,41,42,87,88,89,90,93,94,97,21,22,95,96,60,61,62,63,66,67,68,69,71,73,74,75,76,77,79,80,64,65,91,92,104,105,106,107,110,111,112,114,115,117,118,119,120,123,124,125,126,127,128,129,130,131,140,141,142,143,144,148,149,151,152,153,145,147,113,116,108,109,132,133],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\utils.py":[1,2,3,4,5,6,9,10,13,25,31,39,40,43,46,62,100,123,135,152,162,41,11,26,28,137,139,44,141,143,145,147,164,166,148,14,16,18,20,22,63,65,67,69,71,74,77,79,82,84,87,89,91,92,93,94,88,83,68,64,47,48,50,51,53,55,57,58,59,80,95,96,97,78,32,34,124,126,128,130,132,33,125,35,129,153,155,157,158,165,49,54,52,27,29,56],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\user_models.py":[1,2,5,6,7,16,26,36,47,54,61,68,8,9,10,11,12,13,14,38,39,40,41,42,43,44,49,50,51,52,63,64,65,66,18,19,20,22,23,24,70,71],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\database_models.py":[1,2,3,4,7,8,13,53,64,71,77,9,10,11,73,74,66,67,56,57,60,61,68,69,75,44,46,47,48,49,80,81,82,84,85,86,87],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\product_models.py":[1,4,5,7,15,31,52,59,66,75,82,88,95,102,110,121,8,9,10,11,12,13,104,105,106,107,108,17,18,22,23,24,25,26,27,112,113,114,115,116,118,119,122,123,124,125,126,33,34,35,40,41,42,43,44,45,46,47,68,69,70,72],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\sales_models.py":[1,2,5,6,8,17,23,31,41,9,10,11,12,13,14,15,19,20,21,25,26,27,28,29,33,34,35,36,38,39],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\views\\users_views.py":[1,2,3,4,5,6,8,9,12,13,42,57,58,82,83,95,117,98,99,100,101,103,105,106,107,108,109,110,112,113,16,17,18,19,23,24,25,27,28,29,34,35,36,37,38,39,61,62,63,64,68,69,71,72,74,75,76,78,79,30,85,86,87,88,91,92,26,31,32,33,65,66,67,20,21,22]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\__init__.py":[2,3,4,5,6,7,9,11,12,13,14,15,18,25,33,40],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\views\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\views\\product_views.py":[1,2,3,4,5,6,8,9,10,11,14,15,36,37,47,48,58,59,69,70,97,98,113,114,152,153,172,173,183,190,17,18,19,20,23,24,25,27,28,30,31,32,33,51,52,54,55,62,63,65,66,29,40,41,42,43,100,101,102,103,106,107,110,21,22,108,109,73,74,75,76,79,80,81,82,84,86,87,88,89,90,92,93,77,78,104,105,117,118,119,120,123,124,125,127,128,130,131,132,133,136,137,139,140,141,142,143,144,145,146,147,156,157,158,159,160,164,165,167,168,169,161,163,126,129,121,122,148,149],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\utils.py":[1,2,3,4,5,6,9,10,13,26,32,40,41,44,47,68,109,137,149,167,178,42,11,27,29,151,153,45,155,157,159,161,181,182,184,162,14,16,18,20,22,69,71,73,75,77,80,83,85,88,90,94,97,100,101,102,103,95,96,89,74,70,104,105,106,107,48,49,52,53,56,59,63,64,65,86,84,33,35,138,140,142,144,146,34,139,36,143,168,170,172,173,183,50,51,57,58,54,55,28,30,60,61,62],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\__init__.py":[1],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\user_models.py":[1,2,5,6,7,16,28,39,52,59,66,73,80,89,98,8,9,10,11,12,13,14,41,42,43,44,45,46,47,48,49,83,84,85,86,87,54,55,56,57,18,19,21,23,24,25,26,100,101,68,69,70,71,92,93,94,95,96],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\database_models.py":[1,2,3,4,7,8,13,53,64,71,78,9,10,11,73,74,66,67,56,57,60,61,68,69,75,76,44,46,47,48,49,81,82,83,85,86,87,88],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\product_models.py":[1,4,5,7,16,32,53,60,67,76,83,89,96,103,111,123,133,8,9,10,11,12,13,14,105,106,107,108,109,18,19,23,24,25,26,27,28,114,115,116,117,118,120,121,125,126,127,128,130,131,134,135,136,137,138,34,35,36,41,42,43,44,45,46,47,48,69,70,71,73],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\models\\sales_models.py":[1,2,3,4,7,8,10,19,26,32,41,11,12,13,14,15,16,17,21,22,24,28,29,30,34,35,36,38,39],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\views\\users_views.py":[1,2,3,4,6,7,9,10,13,14,44,45,71,72,87,88,101,102,116,140,119,120,121,122,124,126,127,128,129,130,131,133,134,17,18,19,20,24,25,26,28,29,30,36,37,38,39,40,41,48,49,50,51,55,56,58,59,61,62,63,65,67,68,31,105,106,107,108,111,112,113,75,76,77,78,82,83,84,79,80,81,66,27,32,33,35,91,92,93,94,96,97,98,52,53,54,21,22,23,137],"C:\\Users\\FRED\\PycharmProjects\\StoreManager-ADCv3\\app\\api\\v2\\views\\sales_views.py":[]}}
