+++ app/__init__.py
@@ -5,7 +5,6 @@ from flask_jwt_extended import JWTManager
 from flask_cors import CORS
 from .api.v2.views.product_views import prod as product_bp
 from .api.v2.views.users_views import user_dec as user_bp
-from .api.v2.views.sales_views import sale as sales_bp
 # create our flask application
 flask_app = Flask(__name__)
 # register the blueprints of our models
@@ -14,7 +13,6 @@ flask_app.config['JWT_SECRET_KEY'] = 'hello-there-im-allan'
 jwt = JWTManager(flask_app)
 flask_app.register_blueprint(product_bp)
 flask_app.register_blueprint(user_bp)
-flask_app.register_blueprint(sales_bp)
 
 
 @flask_app.errorhandler(404)
