+++ db.ini
@@ -1,6 +1,6 @@
 [postgresql]
-host=ec2-23-23-153-145.compute-1.amazonaws.com
-database=d1f6m1j1j5tka2
-user=ncvaqdjqyeedvp
+host=ec2-54-235-193-0.compute-1.amazonaws.com
+database=d7hssh64bibdfe
+user=aicqxaigacowfd
 port=5432
-password=2f24a20ca4064f5626dabf9fe1174ae72d89450b6ce6ba6e159271c33b66b2f9
+password=9eec6dda5895a5ace3a2b0d000c123095bd81c2d4e2795b7e2eb1ef8148b4cb5
